+++ pages/index.tsx
@@ -1,12 +1,12 @@
-import * as React from 'react';
-import {connect} from 'react-redux';
-import Layout from '../components/_Layout/Layout';
-import * as blogActions from '../store/blog/actions';
-import * as portfolioActions from '../store/portfolio/actions';
+import * as React from "react";
+import { connect } from "react-redux";
+import Layout from "../components/_Layout/Layout";
+import * as blogActions from "../store/blog/actions";
+import * as portfolioActions from "../store/portfolio/actions";
 
-import {Blog} from '../server/entity/MyBlog';
-import {UserState} from '../store/user/reducers';
-import Hl from '../components/util/Hl';
+import { Blog } from "../server/entity/MyBlog";
+import { UserState } from "../store/user/reducers";
+import Hl from "../components/util/Hl";
 
 interface Props {
   getBlogList: (token?) => void;
@@ -20,8 +20,8 @@ interface State {}
 
 export class IndexPage extends React.Component<Props, State> {
   componentDidMount() {
-    console.log('component mounted');
-    const {token} = this.props.userSession;
+    console.log("component mounted");
+    const { token } = this.props.userSession;
     this.props.getBlogList(token);
     this.props.getPortfolioList();
   }
@@ -31,16 +31,15 @@ export class IndexPage extends React.Component<Props, State> {
       <Layout displayImg showFooter>
         <div className="home-body animated bounceInDown">
           <h1>
-            Hi! I'm <Hl>Alex</Hl> a front end developer with full stack
-            knowledge
+            Hi! I'm <Hl>Alex</Hl> a Frontend developer with full stack knowledge
           </h1>
           <h3>
             Located in <Hl>Singapore</Hl> I have experience with everything from
             Android development to SCSS. I started in a more backend role with
             Java and C# and spent two years working with <Hl>React JS </Hl> in
-            media streaming. Now I am working in a more full stack capacity on a{' '}
+            media streaming. Now I am working in a more full stack capacity on a{" "}
             <Hl>Microsoft Stack</Hl> with a front end focus. In my freetime I
-            enjoy playing with other technologies like <Hl>Rust</Hl> and{' '}
+            enjoy playing with other technologies like <Hl>Rust</Hl> and{" "}
             <Hl>Electron </Hl>
             and learning new and better ways to write code.
           </h3>
@@ -80,7 +79,7 @@ export class IndexPage extends React.Component<Props, State> {
   }
 }
 
-const mapStateToProps = ({blog, user}: any) => ({
+const mapStateToProps = ({ blog, user }: any) => ({
   userSession: user,
   blogList: blog.blogList,
   fetching: blog.fetching,